view report screen & refactoring
- added a screen to view all the expenses in the current report
- refactored the expense card logic to a separate component class so it can be called from the home screen and the view report screen
